+++ CHANGELOG.md
@@ -11,10 +11,14 @@
 ### Removed
 
 ### Fixed
-- Fix rules is empty at the first time
-- Fix settings do not saved
 
 ### Security
+
+## [0.3.0] - 2023-07-20
+
+### Fixed
+- Fix rules is empty at the first time
+- Fix settings do not saved
 
 ## [0.2.0] - 2023-07-19
 
@@ -26,6 +30,7 @@
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/Omico/intellij-settings-hero/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/Omico/intellij-settings-hero/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/Omico/intellij-settings-hero/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/Omico/intellij-settings-hero/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/Omico/intellij-settings-hero/commits/v0.1.0